+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Ensure that a protected environment can't be decommissioned (#2376)
 - Don't load all code on agent start (#2343)
 - Allow empty body in else branch for if-else statement (#2375)
+- Fixed export failure with null in dict (#2437)
 
 # Release 2020.4 (2020-09-08)