+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - Allow empty body in else branch for if-else statement (#2375)
 - Fixed export failure with null in dict (#2437)
 - Fixed small documentation issues
+- Only store single agent instance in database for each distinct instance to prevent database overflow when agent rapidly and
+    repeatadly dis- and reconnects (#2394)
 
 # Release 2020.4 (2020-09-08)